+++ django_ufilter/filtersets/base.py
@@ -122,7 +122,7 @@ class FilterSet(six.with_metaclass(FilterSetMeta, BaseFilter)):
         object from the Django view.
     strict_mode : str, optional
         Strict mode how :class:`.FilterSet` should behave when any validation
-        fails. See :class:`url_filter.constants.StrictMode` doc for more information.
+        fails. See :class:`django_ufilter.constants.StrictMode` doc for more information.
         Default is ``empty``.
     """