+++ test_project/settings.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Bare ``settings.py`` for running tests for url_filter
+# Bare ``settings.py`` for running tests for django_ufilter
 import os
 
 DEBUG = True
@@ -18,7 +18,10 @@ if os.environ.get("USE_POSTGRES") == "True":
     }
 else:
     DATABASES = {
-        "default": {"ENGINE": "django.db.backends.sqlite3", "NAME": "url_filter.sqlite"}
+        "default": {
+            "ENGINE": "django.db.backends.sqlite3",
+            "NAME": "django_ufilter.sqlite",
+        }
     }
 
 INSTALLED_APPS = (
@@ -26,7 +29,7 @@ INSTALLED_APPS = (
     "test_project.many_to_many",
     "test_project.many_to_one",
     "test_project.one_to_one",
-    "url_filter",
+    "django_ufilter",
     "debug_toolbar",
     "django_extensions",
     "rest_framework",
@@ -49,7 +52,7 @@ TEMPLATES = [
 ROOT_URLCONF = "test_project.urls"
 
 REST_FRAMEWORK = {
-    "DEFAULT_FILTER_BACKENDS": ["url_filter.integrations.drf.DjangoFilterBackend"]
+    "DEFAULT_FILTER_BACKENDS": ["django_ufilter.integrations.drf.DjangoFilterBackend"]
 }
 
 DEBUG_TOOLBAR_PANELS = [