+++ tests/filtersets/test_base.py
@@ -5,13 +5,13 @@ import pytest
 from django import forms
 from django.http import QueryDict
 
+from django_ufilter.backends.django import DjangoFilterBackend
+from django_ufilter.constants import StrictMode
+from django_ufilter.exceptions import Empty
+from django_ufilter.filters import Filter
+from django_ufilter.filtersets.base import FilterSet
+from django_ufilter.utils import FilterSpec
 from test_project.one_to_one.models import Restaurant, Waiter
-from url_filter.backends.django import DjangoFilterBackend
-from url_filter.constants import StrictMode
-from url_filter.exceptions import Empty
-from url_filter.filters import Filter
-from url_filter.filtersets.base import FilterSet
-from url_filter.utils import FilterSpec
 
 
 class TestFilterSet(object):