+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function, unicode_literals
 
-from url_filter.utils import FilterSpec, LookupConfig, SubClassDict, dictify
+from django_ufilter.utils import FilterSpec, LookupConfig, SubClassDict, dictify
 
 
 class TestFilterSpec(object):